Added support for lamdas
Lambdas can now be passed as arguments to the Builder constructor. The problem was that the overloads for the append method were defined and declared at the same points, so passing arguments in certain orders caused the compiler to not find the right overload. Now all overloads for append are declared outside of the class