Added support for Builders
Added support for Builders as arguments of the Builder constructor. Now Builders can be initialized from compounds of not only preds, but also builders, in which case, the preds of such builders are appended to the builder being constructed. Also added some comments to the header for ease of development