Renamed PredBuilder as Token and Builder as Pattern.
I changed those names to match the name given to what these classes represent. So yeah, Token represents Tokens as in Tokenization, and Pattern represents a set of Tokens to be matched, like a regular expression. These classes had those names (PredBuilder and Builder) before because they were meant to be factory classes for the actual Pred (now Token) and Pattern classes, but now I have another plan on this project, so that won't be necessary