Added a constructor for Loop TokenLists and some minor changes
Added the 'loop' function template to create Token Lists that start with a Loop Token. Found that giving only 1 argument to the TokenList constructor caused a compiler error originated in 'append', so I added overloads to append for when only 1 argument is given to it. Corrected some comments to be more descriptive, and deleted the 'For Debuggin' from the size function as it is not only for debugging now